fixed gcc/msvc build warnings
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@3842 406a0c4d-033a-0410-8de8-e80135713968

+++ src/yysphinxql.c
@@ -1932,7 +1932,7 @@ yyreduce:
 
   case 40:
 
-    { yyval.m_sValue.SetSprintf ( "%d", yyvsp[0].m_iValue ); ;}
+    { yyval.m_sValue.SetSprintf ( "%lld", INT64(yyvsp[0].m_iValue) ); ;}
     break;
 
   case 41:
